requirements.txt
Django==4.2
djangorestframework==3.14.0
django-filter==23.5
psycopg2-binary==2.9.9  # Если используете PostgreSQL
